Refactored view components of model into view class
